+++ packages/framework-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @meridianjs/framework-utils
 
+## 1.26.0
+
+### Patch Changes
+
+- @meridianjs/types@1.26.0
+
 ## 1.25.0
 
 ### Patch Changes